# Fix NilClass#try and NilClass#try! performance slowdown in Ruby 2.7+