@kamipo 2021-07-15
Can this change also make Tryable#{try,try!} faster?

https://github.com/rails/rails/blob/c4eb3291e9d7034c3d3dd5c77f79652d96879ca5/activesupport/lib/active_support/core_ext/object/try.rb#L6-L32