@schneems 2021-07-16
@kamipo I looked at it and I’m unsure since checking for the first element of an array to be nil seems slightly slower than checking if the var is nil. When in doubt, benchmark. Unfortunately I’m away from my computer. I’m assuming if it’s faster you would like that to be rolled into this change as well.