@kamipo 2021-07-16
I've benchmarked, definitely faster:

```ruby
class FooCurrent
  def try(method_name = nil, *args, &block)
    if method_name.nil? && block_given?
      if block.arity == 0
        instance_eval(&block)
      else
        yield self
      end
    elsif respond_to?(method_name)
      public_send(method_name, *args, &block)
    end
  end
  ruby2_keywords(:try)

  def try!(method_name = nil, *args, &block)
    if method_name.nil? && block_given?
      if block.arity == 0
        instance_eval(&block)
      else
        yield self
      end
    else
      public_send(method_name, *args, &block)
    end
  end
  ruby2_keywords(:try!)

  def anything
  end
end

class FooUpdated
  def try(*args, &block)
    if args.empty? && block_given?
      if block.arity == 0
        instance_eval(&block)
      else
        yield self
      end
    elsif respond_to?(args.first)
      public_send(*args, &block)
    end
  end
  ruby2_keywords(:try)

  def try!(*args, &block)
    if args.empty? && block_given?
      if block.arity == 0
        instance_eval(&block)
      else
        yield self
      end
    else
      public_send(*args, &block)
    end
  end
  ruby2_keywords(:try!)

  def anything
  end
end

require 'benchmark/ips'


foo_updated = FooUpdated.new
foo_current = FooCurrent.new

Benchmark.ips do |x|
  x.report("updated#try") { foo_updated.try(:anything) }
  x.report("current#try") { foo_current.try(:anything) }
  x.report("updated#try!") { foo_updated.try!(:anything) }
  x.report("current#try!") { foo_current.try!(:anything) }
  x.compare!
end
```

```
% ruby -v
ruby 2.7.3p183 (2021-04-05 revision 6847ee089d) [x86_64-darwin17]

% ruby scratch.rb                                                                                        Warming up --------------------------------------
         updated#try   223.111k i/100ms
         current#try   208.684k i/100ms
        updated#try!   283.388k i/100ms
        current#try!   240.961k i/100ms
Calculating -------------------------------------
         updated#try      2.282M (± 1.9%) i/s -     11.602M in   5.086050s
         current#try      2.135M (± 1.5%) i/s -     10.852M in   5.084306s
        updated#try!      2.766M (± 8.4%) i/s -     13.886M in   5.083024s
        current#try!      2.386M (± 2.7%) i/s -     12.048M in   5.054344s

Comparison:
        updated#try!:  2766490.7 i/s
        current#try!:  2385664.7 i/s - 1.16x  (± 0.00) slower
         updated#try:  2281949.6 i/s - 1.21x  (± 0.00) slower
         current#try:  2134814.4 i/s - 1.30x  (± 0.00) slower

% ruby -v
ruby 3.0.0p0 (2020-12-25 revision 95aff21468) [x86_64-darwin17]

% ruby scratch.rb
Warming up --------------------------------------
         updated#try   225.655k i/100ms
         current#try   209.358k i/100ms
        updated#try!   278.623k i/100ms
        current#try!   234.724k i/100ms
Calculating -------------------------------------
         updated#try      2.246M (± 3.5%) i/s -     11.283M in   5.029722s
         current#try      2.110M (± 1.6%) i/s -     10.677M in   5.062279s
        updated#try!      2.796M (± 2.5%) i/s -     14.210M in   5.085431s
        current#try!      2.419M (± 2.6%) i/s -     12.206M in   5.049067s

Comparison:
        updated#try!:  2796170.8 i/s
        current#try!:  2419329.9 i/s - 1.16x  (± 0.00) slower
         updated#try:  2246330.4 i/s - 1.24x  (± 0.00) slower
         current#try:  2109745.1 i/s - 1.33x  (± 0.00) slower
```